tests/wpt/web-platform-tests/svg/animations/animate-number-calcMode-discrete-keyTimes.html

<!doctype html>
<html>
<meta charset="utf-8">
<title>Test calcMode discrete with from-to animation on numbers. You should see a green 100x100 rect and only PASS messages</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/SVGAnimationTestCase-testharness.js"></script>

<svg>
</svg>

<script>
var rootSVGElement = document.querySelector("svg");
var epsilon = 1.0;

// Setup test document
var rect = createSVGElement("rect");
rect.setAttribute("id", "rect");
rect.setAttribute("x", "100");
rect.setAttribute("width", "100");
rect.setAttribute("height", "100");
rect.setAttribute("fill", "green");
rect.setAttribute("onclick", "executeTest()");

var animate = createSVGElement("animate");
animate.setAttribute("id", "animation");
animate.setAttribute("attributeName", "x");
animate.setAttribute("values", "100;200;300");
animate.setAttribute("begin", "0s");
animate.setAttribute("dur", "3s");
animate.setAttribute("keyTimes", "0;0.5;1");
animate.setAttribute("calcMode", "discrete");
animate.setAttribute("fill", "freeze");
rect.appendChild(animate);
rootSVGElement.appendChild(rect);

// Setup animation test
function sample1() {
    assert_equals(rect.x.animVal.value, 100);
    assert_equals(rect.x.baseVal.value, 100);
}

function sample2() {
    assert_equals(rect.x.animVal.value, 200);
    assert_equals(rect.x.baseVal.value, 100);
}

function sample3() {
    assert_equals(rect.x.animVal.value, 300);
    assert_equals(rect.x.baseVal.value, 100);
}

smil_async_test((t) => {
    const expectedValues = [
        // [animationId, time, sampleCallback]
        ["animation", 0.0, sample1],
        ["animation", 1.499, sample1],
        ["animation", 1.501, sample2],
        ["animation", 2.999, sample2],
        ["animation", 3.001, sample3]
    ];

    runAnimationTest(t, expectedValues);
});

window.clickX = 150;

</script>
